@@ -1,10 +1,14 @@
 {
   "name": "builderbook",
-  "version": "2",
+  "version": "0.0.1",
   "license": "MIT",
+  "engines": {
+    "node": "14.18.1",
+    "yarn": "1.22.5"
+  },
   "scripts": {
     "dev": "nodemon server/server.js --watch server",
-    "build": "NODE_ENV=production next build",
+    "build": "next build",
     "start": "node server/server.js",
     "lint": "eslint '**/*.js' '**/*.jsx'",
     "test": "jest --coverage"
@@ -13,44 +17,46 @@
     "coverageDirectory": "./.coverage"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/styles": "^4.11.4",
+    "@emotion/cache": "^11.6.0",
+    "@emotion/react": "^11.6.0",
+    "@emotion/server": "^11.4.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/material": "^5.2.0",
     "@octokit/oauth-authorization-url": "^4.1.0",
     "@octokit/rest": "^18.0.3",
-    "@stripe/stripe-js": "^1.9.0",
-    "aws-sdk": "2.737.0",
-    "compression": "1.7.4",
-    "connect-mongo": "3.2.0",
-    "dotenv": "8.2.0",
-    "express": "4.17.1",
-    "express-session": "1.17.1",
-    "front-matter": "4.0.2",
-    "handlebars": "4.7.6",
-    "he": "1.2.0",
-    "helmet": "4.1.0-rc.2",
-    "highlight.js": "10.1.2",
-    "htmlescape": "1.1.1",
+    "@stripe/stripe-js": "^1.21.1",
+    "aws-sdk": "^2.796.0",
+    "compression": "^1.7.4",
+    "connect-mongo": "^3.2.0",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "express-session": "^1.17.1",
+    "front-matter": "^4.0.2",
+    "he": "^1.2.0",
+    "helmet": "^4.1.0-rc.2",
+    "highlight.js": "^10.1.2",
     "isomorphic-unfetch": "^3.0.0",
-    "lodash": "4.17.20",
-    "marked": "1.1.1",
-    "mongoose": "5.10.0",
-    "next": "^10.2.3",
-    "node-fetch": "^2.6.0",
-    "nprogress": "0.2.0",
-    "passport": "0.4.1",
-    "passport-google-oauth": "2.0.0",
-    "prop-types": "15.7.2",
-    "qs": "6.9.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "sitemap": "6.2.0",
-    "stripe": "8.89.0",
-    "winston": "3.3.3"
+    "lodash": "^4.17.21",
+    "marked": "^1.2.7",
+    "mongoose": "^5.10.0",
+    "next": "^12.0.3",
+    "node-fetch": "^2.6.1",
+    "nprogress": "^0.2.0",
+    "passport": "^0.4.1",
+    "passport-google-oauth": "^2.0.0",
+    "prop-types": "^15.7.2",
+    "qs": "^6.10.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "sitemap": "^6.2.0",
+    "stripe": "^8.129.0",
+    "winston": "^3.3.3"
   },
   "devDependencies": {
+    "@babel/core": "^7.16.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
+    "eslint": "^7.25.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-import": "^2.22.1",
@@ -59,7 +65,7 @@
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^26.6.3",
-    "nodemon": "2.0.6",
+    "nodemon": "^2.0.15",
     "prettier": "^2.2.1"
   }
 }
