@@ -1,32 +1,46 @@
 /* eslint-disable react/no-danger */
 import React from 'react';
 import Document, { Head, Html, Main, NextScript } from 'next/document';
-import { ServerStyleSheets } from '@material-ui/styles';
-// import htmlescape from 'htmlescape';
 
-// const { StripePublishableKey } = process.env;
-// // console.log(StripePublishableKey);
-
-// const env = { StripePublishableKey };
-// // console.log(env);
+import createEmotionServer from '@emotion/server/create-instance';
+import createCache from '@emotion/cache';
 
 class MyDocument extends Document {
   static getInitialProps = async (ctx) => {
     // Render app and page and get the context of the page with collected side effects.
-    const sheets = new ServerStyleSheets();
     const originalRenderPage = ctx.renderPage;
+
+    // You can consider sharing the same emotion cache between all the SSR requests to speed up performance.
+    // However, be aware that it can have global side effects.
+    const cache = createCache({
+      key: 'css',
+      prepend: true,
+    });
+    const { extractCriticalToChunks } = createEmotionServer(cache);
 
     ctx.renderPage = () =>
       originalRenderPage({
-        enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
+        // eslint-disable-next-line react/display-name
+        enhanceApp: (App) => (props) => <App emotionCache={cache} {...props} />,
       });
 
     const initialProps = await Document.getInitialProps(ctx);
+    // This is important. It prevents emotion to render invalid HTML.
+    // See https://github.com/mui-org/material-ui/issues/26561#issuecomment-855286153
+    const emotionStyles = extractCriticalToChunks(initialProps.html);
+    const emotionStyleTags = emotionStyles.styles.map((style) => (
+      <style
+        data-emotion={`${style.key} ${style.ids.join(' ')}`}
+        key={style.key}
+        // eslint-disable-next-line react/no-danger
+        dangerouslySetInnerHTML={{ __html: style.css }}
+      />
+    ));
 
     return {
       ...initialProps,
       // Styles fragment is rendered after the app and page rendering finish.
-      styles: [...React.Children.toArray(initialProps.styles), sheets.getStyleElement()],
+      styles: [...React.Children.toArray(initialProps.styles), ...emotionStyleTags],
     };
   };
 
@@ -37,12 +51,14 @@ class MyDocument extends Document {
           <meta charSet="utf-8" />
           <meta name="google" content="notranslate" />
           <meta name="theme-color" content="#1976D2" />
+
           <link
             rel="shortcut icon"
             href="https://storage.googleapis.com/builderbook/favicon32.png"
           />
 
           <link rel="stylesheet" href="https://storage.googleapis.com/builderbook/vs.min.css" />
+
           <style>
             {`
               a {
@@ -73,7 +89,7 @@ class MyDocument extends Document {
           </style>
           <script
             async
-            src={`https://www.googletagmanager.com/gtag/js?id=${process.env.GA_MEASUREMENT_ID}`}
+            src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID}`}
           />
           <script
             dangerouslySetInnerHTML={{
@@ -81,7 +97,7 @@ class MyDocument extends Document {
               window.dataLayer = window.dataLayer || [];
               function gtag(){dataLayer.push(arguments);}
               gtag('js', new Date());
-              gtag('config', '${process.env.GA_MEASUREMENT_ID}', {
+              gtag('config', '${process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID}', {
                 page_path: window.location.pathname,
               });
             `,
@@ -90,60 +106,11 @@ class MyDocument extends Document {
         </Head>
         <body>
           <Main />
-          {/* Added next.config.js
-          {/* eslint-disable-next-line react/no-danger */}
-          {/* <script dangerouslySetInnerHTML={{ __html: `__ENV__ = ${htmlescape(env)}` }} /> */}
           <NextScript />
         </body>
       </Html>
     );
   }
 }
-
-// MyDocument.getInitialProps = async (ctx) => {
-//   // Resolution order
-//   //
-//   // On the server:
-//   // 1. app.getInitialProps
-//   // 2. page.getInitialProps
-//   // 3. document.getInitialProps
-//   // 4. app.render
-//   // 5. page.render
-//   // 6. document.render
-//   //
-//   // On the server with error:
-//   // 1. document.getInitialProps
-//   // 2. app.render
-//   // 3. page.render
-//   // 4. document.render
-//   //
-//   // On the client
-//   // 1. app.getInitialProps
-//   // 2. page.getInitialProps
-//   // 3. app.render
-//   // 4. page.render
-
-//   // Render app and page and get the context of the page with collected side effects.
-//   const sheets = new ServerStyleSheets();
-//   const originalRenderPage = ctx.renderPage;
-
-//   ctx.renderPage = () =>
-//     originalRenderPage({
-//       enhanceApp: (App) => (props) => sheets.collect(<App {...props} />),
-//     });
-
-//   const initialProps = await Document.getInitialProps(ctx);
-
-//   return {
-//     ...initialProps,
-//     // Styles fragment is rendered after the app and page rendering finish.
-//     styles: (
-//       <React.Fragment>
-//         {initialProps.styles}
-//         {sheets.getStyleElement()}
-//       </React.Fragment>
-//     ),
-//   };
-// };
 
 export default MyDocument;
