@@ -1,9 +1,7 @@
 const winston = require('winston');
 
-const dev = process.env.NODE_ENV !== 'production';
-
 const logger = winston.createLogger({
-  level: dev ? 'debug' : 'info',
+  level: process.env.NODE_ENV === 'production' ? 'info' : 'debug',
   format: winston.format.combine(winston.format.splat(), winston.format.simple()),
   transports: [new winston.transports.Console()],
 });
